Added aria role and keyboard navigation.
The role-attribute is important when we make "none-interactive"
elements interactive, such as SPANs. This makes it easier for
person using screen readers and such to use the site.
(http://www.w3.org/TR/wai-aria/roles)

By having a tabIndex and support "enter" we can easily navigate
to the upload element, and enter it via the keyboard. This makes
our site more usable.