fix: abnormal concurrency in xhr requests
Although `XMLHttpRequest` can asynchronously send requests, this process is implicit and cannot be awaited through `await`. To control the concurrency of `XMLHttpRequest`, it is necessary to convert `XMLHttpRequest` into an explicit asynchronous task. Modify the task completion status in both `xhr.onload` and `xhr.onerror` to enable `asyncPool` to remove completed `pending` requests through `Promise.race` and initiate the next request.

To minimize code changes, add a `requestTasks` queue in the component to store `XMLHttpRequest` request tasks. Additionally, include an `appendRequestTask` parameter in the `request` function to add tasks to the queue.

When the `concurrencyLimit` property is provided to the component, use `asyncPool` to control the concurrency and initiate requests. Otherwise, sequentially initiate all requests in the `requestTasks` queue.